Implement complex entity instantiation
Consider the following example:

ENTITY base
      a: STRING
END_ENTITY;

ENTITY extended
   SUBTYPE OF (base);
      b: STRING
END_ENTITY;

A simple instantiation of the extended entity looks like this:

A complex instantiation allows each subtype to be instatiated
separately like this:

This changes implements support for complex entity instantiation by
cherry-picking fields across all the partial entities in the list. An
complete instance of every entity in the list is then inserted into the
table. In the example above, an instance of base and an instance of
extended would be inserted into the table.

The cherry-picking process is implemented with the help of a table of
every field name for partial and complete entities.