Implement partial/complex entity export
The implementation required a couple of new features:

1. ToData::to_partial function

This allows the generator to choose between serializing partial and
standard entities. The following entities are equivalent:

 Standard: PLANE('X', #1)
 Partial: (ELEMENTARY_SURFACE('X', #1) PLANE())

2. #[holder(supertype = "type_name")] attribute

This is used by the automatically derived implementation of to_partial
to select which fields should be serialized. Partial entity fields are
serialized if they are not already defined by an existing supertype
field.