Add ability to insert into generated tables
The `Insert` trait is used to insert a holder type into a table.
Extra `Insert` implementations are provided for tuples for the
convenience of inserting complex/partial entity instances. This is
limited to five partial entity insertions for now but can be increased
easily by adding more implementations in the ruststep-derive crate.

The user must ensure when inserting into tables that no value is
already associated with a given ID or else panic ensues.